Добавление прав доступа и и связанного с ними функционала

+++ server/src/models/TaskColumn.js
@@ -1,7 +1,16 @@
 const pool = require('../config/db')
+const User = require('./User')
 
-const createTaskColumn = async ({ title, project_id }) => {
+const createTaskColumn = async ({ title, project_id, userId }) => {
+	console.log('title', title)
+	console.log('project_id', project_id)
+	console.log('userId', userId)
 	try {
+		const role = await User.getUserRole(project_id, userId)
+		if (role !== 'owner') {
+			throw new Error('Нет прав на создание столбца')
+		}
+
 		const res = await pool.query(
 			`INSERT INTO "TaskColumn" (title, project_id)
        VALUES ($1, $2)
@@ -11,12 +20,18 @@ const createTaskColumn = async ({ title, project_id }) => {
 		return res.rows[0]
 	} catch (e) {
 		console.error(e)
-		return null
+		return { success: false, message: e.message }
 	}
 }
 
-const getTaskColumnByProjectId = async project_id => {
+const getTaskColumnByProjectId = async (project_id, user_id) => {
 	try {
+		// Получаем роль 
+		console.log('project_id', project_id)
+		console.log('user_id', user_id)
+		const role = await User.getUserRole(project_id, user_id)
+		console.log('role', role)
+		// Получаем колонки с задачами
 		const res = await pool.query(
 			`SELECT
         tc.task_column_id,
@@ -41,18 +56,35 @@ const getTaskColumnByProjectId = async project_id => {
       ORDER BY tc.task_column_id`,
 			[project_id]
 		)
-		return res.rows
+
+		return {
+			success: true,
+			role, // роль пользователя в проекте
+			columns: res.rows,
+		}
 	} catch (e) {
 		console.error(e)
-		return null
+		return { success: false, message: e.message }
 	}
 }
 
-
-
-const deleteTaskColumnById = async task_column_id => {
+const deleteTaskColumnById = async (task_column_id, user_id) => {
 	try {
-		console.log('task_column_id:', task_column_id)
+		// Получаем project_id для столбца
+		const res = await pool.query(
+			`SELECT project_id FROM "TaskColumn" WHERE task_column_id = $1`,
+			[task_column_id]
+		)
+		if (res.rowCount === 0) {
+			return { success: false, message: 'Столбец не найден' }
+		}
+		const project_id = res.rows[0].project_id
+
+		const role = await User.getUserRole(project_id, user_id)
+		if (role !== 'owner') {
+			throw new Error('Нет прав на удаление столбца')
+		}
+
 		await pool.query(`DELETE FROM "Task" WHERE task_column_id = $1`, [
 			task_column_id,
 		])
@@ -64,29 +96,40 @@ const deleteTaskColumnById = async task_column_id => {
 		return { success: true }
 	} catch (e) {
 		console.error(e)
-		return null
+		return { success: false, message: e.message }
 	}
 }
 
-const updateTaskColumnById = async (task_column_id, title) => {
+const updateTaskColumnById = async (task_column_id, title, user_id) => {
 	try {
+		// Получаем project_id для столбца
+		const res = await pool.query(
+			`SELECT project_id FROM "TaskColumn" WHERE task_column_id = $1`,
+			[task_column_id]
+		)
+		if (res.rowCount === 0) {
+			return { success: false, message: 'Столбец не найден' }
+		}
+		const project_id = res.rows[0].project_id
+
+		const role = await User.getUserRole(project_id, user_id)
+		if (role !== 'owner') {
+			throw new Error('Нет прав на обновление столбца')
+		}
+
 		const query = `
       UPDATE "TaskColumn"
       SET title = $1
       WHERE task_column_id = $2
       RETURNING *
     `
-		const res = await pool.query(query, [title, task_column_id])
-		return { success: true, taskColumn: res.rows[0] }
+		const resUpdate = await pool.query(query, [title, task_column_id])
+		return { success: true, taskColumn: resUpdate.rows[0] }
 	} catch (e) {
 		console.error('Ошибка обновления названия столбца:', e)
-		return { success: false, error: e }
+		return { success: false, message: e.message }
 	}
 }
-
-
-
-
 
 module.exports = {
 	createTaskColumn,
